requirements.txt
django
pytest
pytest-django
pymupdf
python-docx 
fpdf
ollama